+++ benry-cmdopt/test/cmdopt_test.rb
@@ -8,7 +8,7 @@ require 'benry/cmdopt'
 
 
 def new_sample_schema()
-  sc = Benry::Cmdopt::Schema.new
+  sc = Benry::CmdOpt::Schema.new
   sc.add(:help   , "-h, --help"            , "show help message.")
   sc.add(:version, "--version"             , "print version")
   sc.add(:file   , "-f, --file=<FILE>"     , "filename")
@@ -25,13 +25,13 @@ end
 
 
 
-class Benry::Cmdopt::Schema::Test < MiniTest::Test
+class Benry::CmdOpt::Schema::Test < MiniTest::Test
 
 
   describe "#parse()" do
 
     before do
-      @schema = Benry::Cmdopt::Schema.new
+      @schema = Benry::CmdOpt::Schema.new
     end
 
     it "[!qw0ac] parses command option definition string." do
@@ -69,7 +69,7 @@ class Benry::Cmdopt::Schema::Test < MiniTest::Test
       pr = proc {
         tuple = sc.__send__(:parse_optdef, "-i, --indent <WIDTH>")
       }
-      ok {pr}.raise?(Benry::Cmdopt::SchemaError,
+      ok {pr}.raise?(Benry::CmdOpt::SchemaError,
                      "-i, --indent <WIDTH>: invalid option definition (use '=--indent' instead of ' --indent').")
     end
 
@@ -79,7 +79,7 @@ class Benry::Cmdopt::Schema::Test < MiniTest::Test
   describe "#dup()" do
 
     before do
-      @schema = Benry::Cmdopt::Schema.new
+      @schema = Benry::CmdOpt::Schema.new
       @schema.add(:help   , "-h" , "print help")
     end
 
@@ -108,7 +108,7 @@ class Benry::Cmdopt::Schema::Test < MiniTest::Test
   describe "#add()" do
 
     before do
-      @schema = Benry::Cmdopt::Schema.new
+      @schema = Benry::CmdOpt::Schema.new
     end
 
     it "[!7hi2d] takes command option definition string." do
@@ -116,7 +116,7 @@ class Benry::Cmdopt::Schema::Test < MiniTest::Test
       sc.add(:indent, "-i, --indent=<WIDTH>", "print help")
       items = sc.instance_eval { @items }
       ok {items.length} == 1
-      ok {items[0]}.is_a?(Benry::Cmdopt::SchemaItem)
+      ok {items[0]}.is_a?(Benry::CmdOpt::SchemaItem)
       ok {items[0].key} == :indent
       ok {items[0].short} == 'i'
       ok {items[0].long} == 'indent'
@@ -139,14 +139,14 @@ class Benry::Cmdopt::Schema::Test < MiniTest::Test
       sc = @schema
       pr = proc { sc.add(nil, "-i <N>", nil) }
       msg = "add(nil, \"-i <N>\"): long option required when option key (1st arg) not specified."
-      ok {pr}.raise?(Benry::Cmdopt::SchemaError, msg)
+      ok {pr}.raise?(Benry::CmdOpt::SchemaError, msg)
     end
 
     it "[!97sn0] raises SchemaError when ',' is missing between short and long options." do
       sc = @schema
       pr = proc { sc.add(:exec, '-x --exec=ARG', "exec") }
       msg = "add(:exec, \"-x --exec=ARG\"): missing ',' between short option and long options."
-      ok {pr}.raise?(Benry::Cmdopt::SchemaError, msg)
+      ok {pr}.raise?(Benry::CmdOpt::SchemaError, msg)
     end
 
     it "[!yht0v] keeps command option definitions." do
@@ -155,7 +155,7 @@ class Benry::Cmdopt::Schema::Test < MiniTest::Test
                       type: Integer, rexp: /\A\d+\z/, enum: ['2', '4', '8']) {|v| v.to_i }
       items = sc.instance_eval { @items }
       ok {items.length} == 1
-      ok {items[0]}.is_a?(Benry::Cmdopt::SchemaItem)
+      ok {items[0]}.is_a?(Benry::CmdOpt::SchemaItem)
       ok {items[0].key} == :indent
       ok {items[0].short} == 'i'
       ok {items[0].long} == 'indent'
@@ -173,7 +173,7 @@ class Benry::Cmdopt::Schema::Test < MiniTest::Test
       pr = proc {
         sc.add("-i, --indent[=<WIDTH>]", "indent width", nil)
       }
-      ok {pr}.raise?(Benry::Cmdopt::SchemaError,
+      ok {pr}.raise?(Benry::CmdOpt::SchemaError,
                      'add("-i, --indent[=<WIDTH>]"): 1st arg should be a Symbol as an option key.')
     end
 
@@ -186,7 +186,7 @@ class Benry::Cmdopt::Schema::Test < MiniTest::Test
           sc.add(:indent, "-i, --indent[=<WIDTH>]", {type: Array}) # Ruby 3
         end
       }
-      ok {pr}.raise?(Benry::Cmdopt::SchemaError,
+      ok {pr}.raise?(Benry::CmdOpt::SchemaError,
                      'add(:indent, "-i, --indent[=<WIDTH>]"): help message required as 3rd argument.')
     end
 
@@ -195,7 +195,7 @@ class Benry::Cmdopt::Schema::Test < MiniTest::Test
       pr = proc {
         sc.add(:indent, "-i, --indent[=<WIDTH>]", "indent width", type: Array)
       }
-      ok {pr}.raise?(Benry::Cmdopt::SchemaError,
+      ok {pr}.raise?(Benry::CmdOpt::SchemaError,
                      "Array: unregistered type.")
     end
 
@@ -204,7 +204,7 @@ class Benry::Cmdopt::Schema::Test < MiniTest::Test
       pr = proc {
         sc.add(:indent, "-i, --indent", "indent width", type: Integer)
       }
-      ok {pr}.raise?(Benry::Cmdopt::SchemaError,
+      ok {pr}.raise?(Benry::CmdOpt::SchemaError,
                      "Integer: type specified in spite of option has no params.")
     end
 
@@ -213,7 +213,7 @@ class Benry::Cmdopt::Schema::Test < MiniTest::Test
       pr = proc {
         sc.add(:indent, "-x, --indent[=<WIDTH>]", "indent width", rexp: '\A\d+\z')
       }
-      ok {pr}.raise?(Benry::Cmdopt::SchemaError,
+      ok {pr}.raise?(Benry::CmdOpt::SchemaError,
                      '"\\\\A\\\\d+\\\\z": regexp pattern expected.')
     end
 
@@ -222,7 +222,7 @@ class Benry::Cmdopt::Schema::Test < MiniTest::Test
       pr = proc {
         sc.add(:indent, "-i, --indent", "indent width", rexp: /\A\d+\z/)
       }
-      ok {pr}.raise?(Benry::Cmdopt::SchemaError,
+      ok {pr}.raise?(Benry::CmdOpt::SchemaError,
                      '/\A\d+\z/: regexp pattern specified in spite of option has no params.')
     end
 
@@ -233,7 +233,7 @@ class Benry::Cmdopt::Schema::Test < MiniTest::Test
       pr = proc {
         sc.add(:indent, "-i <N>", "indent width", enum: "2,4,8")
       }
-      ok {pr}.raise?(Benry::Cmdopt::SchemaError,
+      ok {pr}.raise?(Benry::CmdOpt::SchemaError,
                      '"2,4,8": array or set expected.')
     end
 
@@ -242,7 +242,7 @@ class Benry::Cmdopt::Schema::Test < MiniTest::Test
       pr = proc {
         sc.add(:indent, "-i", "enable indent", enum: [2, 4, 8])
       }
-      ok {pr}.raise?(Benry::Cmdopt::SchemaError,
+      ok {pr}.raise?(Benry::CmdOpt::SchemaError,
                      "[2, 4, 8]: enum specified in spite of option has no params.")
     end
 
@@ -252,7 +252,7 @@ class Benry::Cmdopt::Schema::Test < MiniTest::Test
   describe '#option_help()' do
 
     before do
-      sc = Benry::Cmdopt::Schema.new
+      sc = Benry::CmdOpt::Schema.new
       sc.add(:help   , "-h, --help"            , "show help message.")
       sc.add(:version, "    --version"         , "print version")
       sc.add(:file   , "-f, --file=<FILE>"     , "filename")
@@ -315,7 +315,7 @@ END
     end
 
     it "[!848rm] supports multi-lines help message." do
-      sc = Benry::Cmdopt::Schema.new
+      sc = Benry::CmdOpt::Schema.new
       sc.add(:mode, "-m, --mode=<MODE>", <<END)
 output mode
   v, verbose: print many output
@@ -338,7 +338,7 @@ END
   describe '#_default_format()' do
 
     before do
-      sc = Benry::Cmdopt::Schema.new
+      sc = Benry::CmdOpt::Schema.new
       sc.add(:help   , "-h, --help"            , "show help message.")
       sc.add(:version, "    --version"         , "print version")
       sc.add(:file   , "-f, --file=<FILE>"     , "filename")
@@ -362,7 +362,7 @@ END
     end
 
     it "[!bmr7d] changes min_with according to options." do
-      sc = Benry::Cmdopt::Schema.new
+      sc = Benry::CmdOpt::Schema.new
       sc.add(:help   , '-h', "help")
       sc.add(:version, '-v', "version")
       ok {sc.__send__(:_default_format, nil, 40)} == "  %-8s : %s"
@@ -383,21 +383,21 @@ END
   describe '#_preferred_option_width()' do
 
     it "[!kl91t] shorten option help min width when only single options which take no arg." do
-      sc = Benry::Cmdopt::Schema.new
+      sc = Benry::CmdOpt::Schema.new
       sc.add(:help   , '-h', "help")
       sc.add(:version, '-v', "version")
       ok {sc.__send__(:_preferred_option_width)} == 8
     end
 
     it "[!0koqb] widen option help min width when any option takes an arg." do
-      sc = Benry::Cmdopt::Schema.new
+      sc = Benry::CmdOpt::Schema.new
       sc.add(:help   , '-h', "help")
       sc.add(:indent , '-i[<N>]', "indent")
       ok {sc.__send__(:_preferred_option_width)} == 14
     end
 
     it "[!kl91t] widen option help min width when long option exists." do
-      sc = Benry::Cmdopt::Schema.new
+      sc = Benry::CmdOpt::Schema.new
       sc.add(:help   , '-h', "help")
       sc.add(:version, '-v, --version', "version")
       ok {sc.__send__(:_preferred_option_width)} == 14
@@ -412,7 +412,7 @@ END
   describe '#to_s()' do
 
     it "[!rrapd] '#to_s' is an alias to '#option_help()'." do
-      schema = Benry::Cmdopt::Schema.new
+      schema = Benry::CmdOpt::Schema.new
       schema.add(:help   , "-h, --help"        , "show help message")
       schema.add(:version, "    --version"     , "print version")
       ok {schema.to_s} == schema.option_help()
@@ -424,7 +424,7 @@ END
   describe '#each_option_help()' do
 
     before do
-      sc = Benry::Cmdopt::Schema.new
+      sc = Benry::CmdOpt::Schema.new
       sc.add(:help, "-h, --help", "show help message")
       sc.add(:version, "    --version", "print version")
       sc.add(:debug  , "-D, --debug"  , nil)
@@ -467,7 +467,7 @@ END
   describe '#each()' do
 
     before do
-      @schema = Benry::Cmdopt::Schema.new
+      @schema = Benry::CmdOpt::Schema.new
       @schema.add(:help   , "-h, --help"   , "help message")
       @schema.add(:version, "    --version", "print version")
     end
@@ -476,8 +476,8 @@ END
       items = []
       @schema.each {|x| items << x }
       ok {items.length} == 2
-      ok {items[0]}.is_a?(Benry::Cmdopt::SchemaItem)
-      ok {items[1]}.is_a?(Benry::Cmdopt::SchemaItem)
+      ok {items[0]}.is_a?(Benry::CmdOpt::SchemaItem)
+      ok {items[1]}.is_a?(Benry::CmdOpt::SchemaItem)
       ok {items[0].key} == :help
       ok {items[1].key} == :version
       keys = @schema.each.collect {|x| x.key }
@@ -535,14 +535,14 @@ end
 
 
 
-class Benry::Cmdopt::SchemaItem::Test < MiniTest::Test
+class Benry::CmdOpt::SchemaItem::Test < MiniTest::Test
 
 
   describe '#validate_and_convert()' do
 
     def new_item(key, optstr, short, long, param, help,
                  optional: nil, type: nil, rexp: nil, enum: nil, value: nil, &callback)
-      return Benry::Cmdopt::SchemaItem.new(key, optstr, short, long, param, help,
+      return Benry::CmdOpt::SchemaItem.new(key, optstr, short, long, param, help,
                  optional: optional, type: type, rexp: rexp, enum: enum, value: value, &callback)
     end
 
@@ -620,13 +620,13 @@ end
 
 
 
-class Benry::Cmdopt::Parser::Test < MiniTest::Test
+class Benry::CmdOpt::Parser::Test < MiniTest::Test
 
 
   describe '#parse_options()' do
 
     before do
-      @parser = Benry::Cmdopt::Parser.new(new_sample_schema())
+      @parser = Benry::CmdOpt::Parser.new(new_sample_schema())
     end
 
     it "[!3wmsy] returns command option values as a dict." do
@@ -691,13 +691,13 @@ class Benry::Cmdopt::Parser::Test < MiniTest::Test
         errcls = err.class
       }
       ok {errmsg} == "-ix: integer expected."
-      ok {errcls} == Benry::Cmdopt::OptionError
+      ok {errcls} == Benry::CmdOpt::OptionError
       #
-      sc = Benry::Cmdopt::Schema.new
+      sc = Benry::CmdOpt::Schema.new
       sc.add(:file, "--file=<FILE>", "file") do |val|
         File.open(val) {|f| f.read }
       end
-      parser = Benry::Cmdopt::Parser.new(sc)
+      parser = Benry::CmdOpt::Parser.new(sc)
       pr = proc { parser.parse(["--file=/foo/bar/baz.png"]) }
       ok {pr}.raise?(Errno::ENOENT, /No such file or directory/)
     end
@@ -713,41 +713,41 @@ class Benry::Cmdopt::Parser::Test < MiniTest::Test
   describe '#parse_long_option()' do
 
     before do
-      @parser = Benry::Cmdopt::Parser.new(new_sample_schema())
+      @parser = Benry::CmdOpt::Parser.new(new_sample_schema())
     end
 
     it "[!3i994] raises OptionError when invalid long option format." do
       argv = ["--f/o/o"]
       pr = proc { @parser.parse(argv) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError, "--f/o/o: invalid long option.")
+      ok {pr}.raise?(Benry::CmdOpt::OptionError, "--f/o/o: invalid long option.")
     end
 
     it "[!er7h4] raises OptionError when unknown long option." do
       argv = ["--foo"]
       pr = proc { @parser.parse(argv) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError, "--foo: unknown long option.")
+      ok {pr}.raise?(Benry::CmdOpt::OptionError, "--foo: unknown long option.")
     end
 
     it "[!2jd9w] raises OptionError when no arguments specified for arg required long option." do
       argv = ["--file"]
       pr = proc { @parser.parse(argv) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError, "--file: argument required.")
+      ok {pr}.raise?(Benry::CmdOpt::OptionError, "--file: argument required.")
     end
 
     it "[!qyq8n] raises optionError when an argument specified for no arg long option." do
       argv = ["--version=1.0.0"]
       pr = proc { @parser.parse(argv) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError, "--version=1.0.0: unexpected argument.")
+      ok {pr}.raise?(Benry::CmdOpt::OptionError, "--version=1.0.0: unexpected argument.")
     end
 
     it "[!o596x] validates argument value." do
       argv = ["--indent=abc"]
       pr = proc { @parser.parse(argv) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError, "--indent=abc: integer expected.")
+      ok {pr}.raise?(Benry::CmdOpt::OptionError, "--indent=abc: integer expected.")
       #
       argv = ["--path=/foo/bar"]
       pr = proc { @parser.parse(argv) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError, "--path=/foo/bar: directory not exist.")
+      ok {pr}.raise?(Benry::CmdOpt::OptionError, "--path=/foo/bar: directory not exist.")
 
     end
 
@@ -757,19 +757,19 @@ class Benry::Cmdopt::Parser::Test < MiniTest::Test
   describe '#parse_short_option()' do
 
     before do
-      @parser = Benry::Cmdopt::Parser.new(new_sample_schema())
+      @parser = Benry::CmdOpt::Parser.new(new_sample_schema())
     end
 
     it "[!4eh49] raises OptionError when unknown short option specified." do
       argv = ["-hxf", "foo.png"]
       pr = proc { @parser.parse(argv) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError, "-x: unknown option.")
+      ok {pr}.raise?(Benry::CmdOpt::OptionError, "-x: unknown option.")
     end
 
     it "[!utdbf] raises OptionError when argument required but not specified." do
       argv = ["-hf"]
       pr = proc { @parser.parse(argv) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError, "-f: argument required.")
+      ok {pr}.raise?(Benry::CmdOpt::OptionError, "-f: argument required.")
     end
 
     it "[!f63hf] short option arg can be specified without space separator." do
@@ -796,11 +796,11 @@ class Benry::Cmdopt::Parser::Test < MiniTest::Test
     it "[!yu0kc] validates short option argument." do
       argv = ["-iaaa"]
       pr = proc { @parser.parse(argv) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError, "-iaaa: integer expected.")
+      ok {pr}.raise?(Benry::CmdOpt::OptionError, "-iaaa: integer expected.")
       #
       argv = ["-I", "/foo/bar"]
       pr = proc { @parser.parse(argv) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError, "-I /foo/bar: directory not exist.")
+      ok {pr}.raise?(Benry::CmdOpt::OptionError, "-I /foo/bar: directory not exist.")
     end
 
   end
@@ -809,7 +809,7 @@ class Benry::Cmdopt::Parser::Test < MiniTest::Test
   describe '#new_options_dict()' do
 
     it "[!vm6h0] returns new hash object." do
-      parser = Benry::Cmdopt::Parser.new(new_sample_schema())
+      parser = Benry::CmdOpt::Parser.new(new_sample_schema())
       ret = parser.__send__(:new_options_dict)
       ok {ret}.is_a?(Hash)
       ok {ret} == {}
@@ -822,15 +822,15 @@ end
 
 
 
-class Benry::Cmdopt::Test < MiniTest::Test
+class Benry::CmdOpt::Test < MiniTest::Test
 
 
   describe 'PARAM_TYPES[Integer]' do
 
     before do
-      sc = Benry::Cmdopt::Schema.new
+      sc = Benry::CmdOpt::Schema.new
       sc.add(:indent, "-i, --indent[=<N>]", "indent width", type: Integer)
-      @parser = Benry::Cmdopt::Parser.new(sc)
+      @parser = Benry::CmdOpt::Parser.new(sc)
     end
 
     it "[!6t8cs] converts value into integer." do
@@ -843,10 +843,10 @@ class Benry::Cmdopt::Test < MiniTest::Test
 
     it "[!nzwc9] raises error when failed to convert value into integer." do
       pr = proc { @parser.parse(['-i2.1']) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError, "-i2.1: integer expected.")
+      ok {pr}.raise?(Benry::CmdOpt::OptionError, "-i2.1: integer expected.")
       #
       pr = proc { @parser.parse(['--indent=2.2']) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError, "--indent=2.2: integer expected.")
+      ok {pr}.raise?(Benry::CmdOpt::OptionError, "--indent=2.2: integer expected.")
     end
 
   end
@@ -855,9 +855,9 @@ class Benry::Cmdopt::Test < MiniTest::Test
   describe 'PARAM_TYPES[Float]' do
 
     before do
-      sc = Benry::Cmdopt::Schema.new
+      sc = Benry::CmdOpt::Schema.new
       sc.add(:ratio, "-r, --ratio=<RATIO>", "ratio", type: Float)
-      @parser = Benry::Cmdopt::Parser.new(sc)
+      @parser = Benry::CmdOpt::Parser.new(sc)
     end
 
     it "[!gggy6] converts value into float." do
@@ -870,10 +870,10 @@ class Benry::Cmdopt::Test < MiniTest::Test
 
     it "[!t4elj] raises error when faield to convert value into float." do
       pr = proc { @parser.parse(['-r', 'abc']) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError, "-r abc: float expected.")
+      ok {pr}.raise?(Benry::CmdOpt::OptionError, "-r abc: float expected.")
       #
       pr = proc { @parser.parse(['--ratio=abc']) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError, "--ratio=abc: float expected.")
+      ok {pr}.raise?(Benry::CmdOpt::OptionError, "--ratio=abc: float expected.")
     end
 
   end
@@ -882,9 +882,9 @@ class Benry::Cmdopt::Test < MiniTest::Test
   describe 'PARAM_TYPES[TrueClass]' do
 
     before do
-      sc = Benry::Cmdopt::Schema.new
+      sc = Benry::CmdOpt::Schema.new
       sc.add(:border  , "-b, --border[=<on|off>]", "enable border", type: TrueClass)
-      @parser = Benry::Cmdopt::Parser.new(sc)
+      @parser = Benry::CmdOpt::Parser.new(sc)
     end
 
     it "[!47kx4] converts 'true'/'on'/'yes' into true." do
@@ -921,11 +921,11 @@ class Benry::Cmdopt::Test < MiniTest::Test
 
     it "[!h8ayh] raises error when failed to convert value into true nor false." do
       pr = proc { @parser.parse(["-bt"]) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError,
+      ok {pr}.raise?(Benry::CmdOpt::OptionError,
                      "-bt: boolean expected.")
       #
       pr = proc { @parser.parse(["--border=t"]) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError,
+      ok {pr}.raise?(Benry::CmdOpt::OptionError,
                      "--border=t: boolean expected.")
     end
 
@@ -936,9 +936,9 @@ class Benry::Cmdopt::Test < MiniTest::Test
 
     before do
       require 'date'
-      sc = Benry::Cmdopt::Schema.new
+      sc = Benry::CmdOpt::Schema.new
       sc.add(:date, "-d, --date=<YYYY-MM-DD>]", "date", type: Date)
-      @parser = Benry::Cmdopt::Parser.new(sc)
+      @parser = Benry::CmdOpt::Parser.new(sc)
     end
 
     it "[!sru5j] converts 'YYYY-MM-DD' into date object." do
@@ -951,21 +951,21 @@ class Benry::Cmdopt::Test < MiniTest::Test
 
     it "[!h9q9y] raises error when failed to convert into date object." do
       pr = proc { @parser.parse(['-d', '2000/01/01']) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError,
+      ok {pr}.raise?(Benry::CmdOpt::OptionError,
                      "-d 2000/01/01: invalid date format (ex: '2000-01-01')")
       #
       pr = proc { @parser.parse(['--date=01-01-2000']) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError,
+      ok {pr}.raise?(Benry::CmdOpt::OptionError,
                      "--date=01-01-2000: invalid date format (ex: '2000-01-01')")
     end
 
     it "[!i4ui8] raises error when specified date not exist." do
       pr = proc { @parser.parse(['-d', '2001-02-29']) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError,
+      ok {pr}.raise?(Benry::CmdOpt::OptionError,
                      "-d 2001-02-29: date not exist.")
       #
       pr = proc { @parser.parse(['--date=2001-02-29']) }
-      ok {pr}.raise?(Benry::Cmdopt::OptionError,
+      ok {pr}.raise?(Benry::CmdOpt::OptionError,
                      "--date=2001-02-29: date not exist.")
     end
 
@@ -975,8 +975,8 @@ class Benry::Cmdopt::Test < MiniTest::Test
   describe '.new()' do
 
     it "[!7kkqv] creates Facade object." do
-      obj = Benry::Cmdopt.new
-      ok {obj}.is_a?(Benry::Cmdopt::Facade)
+      obj = Benry::CmdOpt.new
+      ok {obj}.is_a?(Benry::CmdOpt::Facade)
     end
 
   end
@@ -986,13 +986,13 @@ end
 
 
 
-class Benry::Cmdopt::Facade::Test < MiniTest::Test
+class Benry::CmdOpt::Facade::Test < MiniTest::Test
 
 
   describe "#dup()" do
 
     before do
-      @cmdopt = Benry::Cmdopt.new
+      @cmdopt = Benry::CmdOpt.new
       @cmdopt.add(:help   , "-h" , "help message")
     end
 
@@ -1020,7 +1020,7 @@ class Benry::Cmdopt::Facade::Test < MiniTest::Test
   describe '#add()' do
 
     it "[!vmb3r] defines command option." do
-      cmdopt = Benry::Cmdopt.new()
+      cmdopt = Benry::CmdOpt.new()
       cmdopt.add(:help, "-h, --help", "show help message")
       items = cmdopt.instance_eval { @schema.instance_variable_get('@items') }
       ok {items}.is_a?(Array)
@@ -1037,7 +1037,7 @@ class Benry::Cmdopt::Facade::Test < MiniTest::Test
   describe '#option_help()' do
 
     before do
-      @cmdopt = Benry::Cmdopt.new
+      @cmdopt = Benry::CmdOpt.new
       @cmdopt.add(:help   , "-h, --help"        , "show help message")
       @cmdopt.add(:version, "    --version"     , "print version")
       @cmdopt.add(:file   , "-f, --file=<FILE>" , "filename")
@@ -1058,7 +1058,7 @@ END
   describe '#to_s()' do
 
     it "[!s61vo] '#to_s' is an alias to '#option_help()'." do
-      cmdopt = Benry::Cmdopt.new
+      cmdopt = Benry::CmdOpt.new
       cmdopt.add(:help   , "-h, --help"        , "show help message")
       cmdopt.add(:version, "    --version"     , "print version")
       ok {cmdopt.to_s} == cmdopt.option_help()
@@ -1070,7 +1070,7 @@ END
   describe '#each_option_help()' do
 
     before do
-      @cmdopt = Benry::Cmdopt.new
+      @cmdopt = Benry::CmdOpt.new
       @cmdopt.add(:help   , "-h, --help"        , "show help message")
       @cmdopt.add(:version, "    --version"     , "print version")
       @cmdopt.add(:file   , "-f, --file=<FILE>" , "filename")
@@ -1095,7 +1095,7 @@ END
   describe '#each()' do
 
     before do
-      @cmdopt = Benry::Cmdopt.new
+      @cmdopt = Benry::CmdOpt.new
       @cmdopt.add(:help   , "-h, --help"   , "help message")
       @cmdopt.add(:version, "    --version", "print version")
     end
@@ -1104,8 +1104,8 @@ END
       items = []
       @cmdopt.each {|x| items << x }
       ok {items.length} == 2
-      ok {items[0]}.is_a?(Benry::Cmdopt::SchemaItem)
-      ok {items[1]}.is_a?(Benry::Cmdopt::SchemaItem)
+      ok {items[0]}.is_a?(Benry::CmdOpt::SchemaItem)
+      ok {items[1]}.is_a?(Benry::CmdOpt::SchemaItem)
       ok {items[0].key} == :help
       ok {items[1].key} == :version
       keys = @cmdopt.each.collect {|x| x.key }
@@ -1118,7 +1118,7 @@ END
   describe '#parse()' do
 
     before do
-      @cmdopt = Benry::Cmdopt.new()
+      @cmdopt = Benry::CmdOpt.new()
       @cmdopt.add(:file, "-f, --file=<FILE>", "file") do |val|
         File.open(val) {|f| f.read }
       end
@@ -1133,7 +1133,7 @@ END
         errcls = err.class
       }
       ok {errmsg} == "-dx: integer expected."
-      ok {errcls} == Benry::Cmdopt::OptionError
+      ok {errcls} == Benry::CmdOpt::OptionError
       #
       pr = proc do
         @cmdopt.parse(["-f", "/foo/bar/baz.png"])